requirements.txt
six~=1.16.0
Faker
pillow
markdown
python-dateutil

django~=4.2.3
django-environ
django-filter~=23.2
django-allauth
django-resized
python-decouple
django-ckeditor
django-crispy-forms
crispy-bootstrap5

django-browser-reload

pytz~=2023.3
DateTime~=5.2
djangorestframework~=3.14.0